+++ backend/src/main/java/org/bihealth/mi/risk_assessment_api/repository/activity/DataSharingActivityRepository.java
@@ -4,7 +4,7 @@ import org.bihealth.mi.risk_assessment_api.model.activity.DataSharingActivity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface DataSharingActivityRepository extends JpaRepository<DataSharingActivity, Integer> {
+public interface DataSharingActivityRepository extends JpaRepository<DataSharingActivity, Long> {
     List<DataSharingActivity> findByCreatorUsername(String creatorUsername);
     List<DataSharingActivity> findBySharedUsernamesContains(String username);
 }
