+++ backend/src/main/java/org/bihealth/mi/risk_assessment_api/repository/dataset/DatasetRepository.java
@@ -4,7 +4,7 @@ import org.bihealth.mi.risk_assessment_api.model.dataset.Dataset;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface DatasetRepository extends JpaRepository<Dataset, Integer> {
+public interface DatasetRepository extends JpaRepository<Dataset, Long> {
     List<Dataset> findByCreatorUsername(String creatorUsername);
     List<Dataset> findBySharedUsernamesContains(String username);
 }
