+++ backend/src/main/java/org/bihealth/mi/risk_assessment_api/repository/dataset/DatasetTableRepository.java
@@ -4,6 +4,6 @@ import org.bihealth.mi.risk_assessment_api.model.dataset.DatasetTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-public interface DatasetTableRepository extends JpaRepository<DatasetTable, Integer> {
-    Optional<DatasetTable> findByIdAndDatasetId(Integer id, Integer datasetId);
+public interface DatasetTableRepository extends JpaRepository<DatasetTable, Long> {
+    Optional<DatasetTable> findByIdAndDatasetId(Long id, Long datasetId);
 }
